This is a rather large change in the code structure. Following some of Vinod's suggestions. I have moved all of the .h files that were previously in subdirectories of "include" into the root "include" directory
moved dlflcore, dlflaux, arcball, vecmat into the main src directory
updated all references to these files, qmake .pro files, etc...
fixed topmod.pro so it will now build exe's into the bin/release and bin/debug folders in the root bin directory OUTSIDE of the src folder.
This is a great start to code reorganization in my opinion. I also began to implement a "tools" folder where I have placed create_torus_tool.cc and create_torus_tool.h from Fenghui's code taken from HighgenusMode.cc
The button has not been moved yet, because this will require some more thought about whether loading a particular geometric object should be considered a "tool". I understand that when creating a torus the user is given quite a few options, but for all the other primitives, the user simply has to click a button, so I wouldn't necessarily consider those buttons "tools". They are in the defined as simple QActions in the MainWindow class right now, but in the future if we want the user to be able to set scaling and other options upon creation of a primitive, perhaps we should put them into separate tool classes so it's easier to change out in the future...